v4.0.4 - 07.02.2018
The asm.js version of the SDK didn't work in IE11 and Edge due to an emscripten issue introduced in 1.37.29.
The WASM SDK is now being outputted using the current 1.37.33 emscripten SDK and the asm.js version of BRFv4 is now being outputted using 1.37.22 of emscripten.
I've submitted a pull request to fix the emscripten issue.